added primitive styling to StockList and renamed Portfolio to StockList
