+++ src/components/StockEntry/StockEntry.css
@@ -1,0 +1,7 @@
+.row {
+  display: flex;
+  flex-direction: row;
+  padding: 1%;
+  border: black 1px;
+  border-style: solid;
+}
