+++ src/components/misc/DisplayBox.js
@@ -1,11 +1,11 @@
 import React from "react";
-
+import "./DisplayBox.css";
 // this should be a simple box that takes info from the backend
 // and displays it
 // add functionality to make increasing values green and decreasing
 // values red
 const DisplayBox = () => {
-  return <div> Display some info here</div>;
+  return <div className="window"> Display some info here</div>;
 };
 
 export default DisplayBox;
